Makefile: Fix build race condition on output path
The $(ARCH) target creates the output directory for binaries. If compile tasks
are started in parallel, this may result in errors like this:
  Fatal error: can't create o-optimize/dyplo_driver.o: No such file or directory

To solve the issue, add a dependency on the $(ARCH) target to make sure the
output directory exists before attempting to write object files.

This fixes intermittent failures of the build.